feat: separate dashboard components and add streaming
